Refactor workout middlewares
