+++ src/types/enums.ts
@@ -12,3 +12,10 @@ export enum WorkoutAccessMode {
   private = "private",
   public = "public",
 }
+
+export enum ScheduleStatus {
+  pending = "pending",
+  accepted = "accepted",
+  amendment = "amendment",
+  declined = "declined",
+}
